requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.2
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
dj-database-url==2.2.0
Django==4.2.7
django-bootstrap-v5==1.0.11
django-filter==23.5
django-heroku==0.3.1
django-jazzmin==3.0.0
django-resized==1.0.2
django-widget-tweaks==1.5.0
filelock==3.14.0
Flask==3.0.3
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
MouseInfo==0.1.3
multidict==6.0.5
packaging==24.0
Pillow==10.1.0
platformdirs==4.2.2
pluggy==1.5.0
psycopg2==2.9.9
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyperclip==1.8.2
pyproject-api==1.6.1
PyRect==0.2.0
PyScreeze==0.1.30
python-decouple==3.8
pytweening==1.2.0
pywhatkit==5.4
requests==2.32.3
sorl-thumbnail==12.10.0
soupsieve==2.5
sqlparse==0.4.4
tox==4.15.0
twilio==9.1.0
typing_extensions==4.12.1
tzdata==2023.3
urllib3==2.2.1
virtualenv==20.26.2
Werkzeug==3.0.3
whitenoise==6.6.0
wikipedia==1.4.0
yarl==1.9.4
